+++ dto/get_userdetails_res.go
@@ -1,9 +1,9 @@
 package dto
 
 type UserDetailsRes struct {
-	Id           int    `json:"id"`
-	Email        string `json:"email"`
-	WalletID     int    `json:"wallet_id"`
-	WalletNumber int    `json:"wallet_number"`
-	Balance      int    `json:"balance"`
+	Id           int     `json:"id"`
+	Email        string  `json:"email"`
+	WalletID     int     `json:"wallet_id"`
+	WalletNumber int     `json:"wallet_number"`
+	Balance      float32 `json:"balance"`
 }
