+++ dto/post_topup_req.go
@@ -1,6 +1,6 @@
 package dto
 
 type TopupReq struct {
-	Amount          int `json:"amount" binding:"required"`
-	SourceOfFundsID int `json:"source_of_funds" binding:"required"`
+	Amount          float32 `json:"amount" binding:"required"`
+	SourceOfFundsID int     `json:"source_of_funds" binding:"required"`
 }
