+++ dto/post_transfer_req.go
@@ -1,7 +1,7 @@
 package dto
 
 type TransferReq struct {
-	ReceiverWalletNumber int    `json:"receiver_wallet_number" binding:"required"`
-	Amount               int    `json:"amount" binding:"required"`
-	Description          string `json:"description" `
+	ReceiverWalletNumber int     `json:"receiver_wallet_number" binding:"required"`
+	Amount               float32 `json:"amount" binding:"required"`
+	Description          string  `json:"description" `
 }
