+++ models/wallet.go
@@ -1,9 +1,9 @@
 package models
 
 type Wallet struct {
-	Id            int `json:"id" gorm:"primarykey"`
-	WalletNumber  int `json:"wallet_number"`
-	DepositNumber int `json:"deposit_number"`
-	UserID        int `json:"user_id"`
-	Balance       int `json:"balance"`
+	Id            int     `json:"id" gorm:"primarykey"`
+	WalletNumber  int     `json:"wallet_number"`
+	DepositNumber int     `json:"deposit_number"`
+	UserID        int     `json:"user_id"`
+	Balance       float32 `json:"balance"`
 }
